Fix up interact piki except the small regswap

+++ include/efx/TPk.h
@@ -19,7 +19,7 @@ void createSimpleWaterOff(Vector3f&);
 void createSimpleGedoku(Vector3f&);
 void createSimpleBlackDrop(Vector3f&);
 void createSimpleDive(Vector3f&);
-void createSimpleChiru(Vector3f&, int);
+void createSimpleChiru(Vector3f&, long);
 
 struct TPkAp : public TSimple2 {
 	inline TPkAp()
